fix(desktop): restrict telemetry model IDs
